+++ crons/HourlyCron.php
@@ -19,6 +19,20 @@ $act = new Accounts();
 
 require_once('../inc/ExchangeAPIs/bitfinex.php');
 
+
+// check that the crons database exists //
+$cronsTableSQL = '
+	CREATE TABLE IF NOT EXISTS `'.$config['db']['prefix'].'CronRuns` (
+	  `id` int(11) NOT NULL AUTO_INCREMENT,
+	  `cron_id` tinyint(1) NOT NULL,
+	  `lastrun` datetime NOT NULL,
+	  `details` varchar(256) NOT NULL,
+	  PRIMARY KEY (`id`)
+	)';
+$rt = $db->iquery($cronsTableSQL);
+
+
+
 // * Get All Active BFX Accounts     * //
 // * Create Account Objects for them * //
 
@@ -27,6 +41,9 @@ foreach($userIds as $uid){
 	$accounts[$uid['id']] = new $act($uid['id']);
 	/* Update their account history */	
 	$accounts[$uid['id']]->bfx->bitfinex_updateHistory();
+	// mark it in the crons table so we know its working
+	echo "INSERT into `".$config['db']['prefix']."CronRuns` (`cron_id`, `lastrun`, `details`) VALUES ('1', NOW(), 'Updated User ".$uid['id']." History')";
+	$cronUpdates = $db->iquery("INSERT into `".$config['db']['prefix']."CronRuns` (`cron_id`, `lastrun`, `details`) VALUES ('1', NOW(), 'Updated User ".$uid['id']." History')");	
 }